Edit exercise2 solution. Create library for stack.
