+++ Lecture_3/Exercise_2/makefile
@@ -1,4 +1,4 @@
-exercise2: exercise2.c mylist.o
+exercise2: exercise2.c mystack.o mylist.o
 	gcc -o $@ $^
 
 mylist.o: ../myLib/mylist.c
